rename component folder to add-problem-type
